Function to give neighbours of cells

TO DO

fN(k, nr, nc, dx = NA)

Arguments

k
cell index (only single values accepted)

nr
number fo rows in original raster

nc
number fo columnc in original raster

Value

the index of neighbouring cells

Details

This is in effect a cut down version of raster adjacent for use with the vectors exctraced by raster::getValues. See example below

Examples

## create a test raster
rst <- raster::raster(matrix(1:200,20,10))
nc <- ncol(rst)
nr <- nrow(rst)
## specify some locations
kvec <- c(1, # top left
          floor(nc/2), # top row middle column
          nc, # top right
          (nc*floor(nr/2))+1, # first column, middle row
          (nc*floor(nr/2))+floor(nc/2), # middle column middle row
          (nc*floor(nr/2))+nc, # last column, middle row
          (nc*(nr-1))+1, # bottom left
          (nc*nr)- floor(nc/2), # bottom row middle
          (nc*nr)) # bottom right

## test matching of locations
for(k in kvec){
    ra <- sort( raster::adjacent(rst,k,direction=8,pairs=FALSE) )
    fn <- sort( fN(k,nrow(rst),ncol(rst)) )
    if( length(ra)!=length(fn) || any( ra!=fn ) ){
        stop(paste("fN does not match raster adjacent for value",k))
    }
}