R6 Class for processing a catchment to make a Dynamic TOPMODEL

This package contains the code for setting up a dynamic TOPMODEL implimentation

Methods

Public methods

dynatopGIS$new()

dynatopGIS$add_layer()

dynatopGIS$get_layer()

dynatopGIS$plot_layer()

dynatopGIS$add_channel()

dynatopGIS$get_channel()

dynatopGIS$sink_fill()

dynatopGIS$compute_properties()

dynatopGIS$classify()

dynatopGIS$get_class()

dynatopGIS$plot_class()

dynatopGIS$create_model()

dynatopGIS$get_model()

dynatopGIS$plot_model()

dynatopGIS$neighbour()

dynatopGIS$get_class_method()

dynatopGIS$get_version()

dynatopGIS$clone()

Method new()

Reads the DEM file, sets up the raster catchment description

Usage

dynatopGIS$new(dem, fill_na = TRUE)

Arguments

dem

a RasterLayer containing the dem

fill_na

should NA values in dem be filled. See details

Details

Reads in the dem using the raster package. If fill_na all NA values other then those that link to the edge of the dem are filled with Inf, so they can be identified as sinks.

Returns

A new `dynatopGIS` object

Method add_layer()

Add a layer of geographical information

Usage

dynatopGIS$add_layer(raster_layer, layer_name, check_name = TRUE)

Arguments

raster_layer

a RasterLayer object to add

layer_name

name to give to the layer

check_name

logical indicating if name should be checked against reserved names

Details

The default is to check the name so that names reserved for computed varaibles are not overwritten. Disabling to overwrite computed layer may have unintended consequences.

Returns

suitable for chaining

Method get_layer()

Get a layer of geographical information or a list of layer names

Usage

dynatopGIS$get_layer(layer_name = character(0))

Arguments

layer_name

name of the layer give to the layer

Returns

a RasterLayer of the requested information if layer_name is given else a vector of layer names

Method plot_layer()

Plot a layer

Usage

dynatopGIS$plot_layer(layer_name, add_channel = TRUE)

Arguments

layer_name

the name of layer to plot

add_channel

shouel the channel be added to the plot

Returns

a plot

Method add_channel()

Import channel data from an OGR file to the `dynatopGIS` object

Usage

dynatopGIS$add_channel(
  sp_object,
  property_names = c(length = "length", startNode = "startNode", endNode = "endNode",
    width = "width"),
  default_width = 2
)

Arguments

sp_object

a spatialLinesDataFrame containing the channel information

property_names

named vector of columns of the spatial data frame to use for channel properties

default_width

default width of a channel if not specified in property_names. Defaults to 2 metres.

Details

Takes the input file and covnerts it to polygons with properties length,width,startNode,endNode. The variable names in the sp_object data frame which corresponding to these properties can be specified in the property_names vector. The function opulates the channel_id & channel_area layers and alters the land_area layer to correspond.

Returns

suitable for chaining

Method get_channel()

Get the channel description

Usage

dynatopGIS$get_channel()

Returns

an SpatialPolygonDataFrame containing the channels

Method sink_fill()

The sink filling algorithm of Planchona and Darboux 2001

Usage

dynatopGIS$sink_fill(
  min_grad = 1e-04,
  max_it = 1e+06,
  verbose = FALSE,
  hot_start = FALSE
)

Arguments

min_grad

Minimum gradient between cell centres

max_it

maximum number of replacement cycles

verbose

print out additional diagnostic information

hot_start

start from filled_dem if it exists

Details

The algorithm implimented in Planchona and Darboux, "A fast, simple and versatile algorithm to fill the depressions in digital elevation models" Catena 46 (2001). A pdf can be found at https://horizon.documentation.ird.fr/exl-doc/pleins_textes/pleins_textes_7/sous_copyright/010031925.pdf.

Method compute_properties()

Computes statistics e.g. log(a/tanb) for raster cells

Usage

dynatopGIS$compute_properties(missing_grad = 1e-04, verbose = FALSE)

Arguments

missing_grad

gradient that can be assigned to a pixel if it can't be computed

verbose

print out additional diagnostic information

Details

The algorithm works in two passes. The first computes the number of upstream pixels and values that do not depend on ordering. The second sequences downslope to compute values. Missing gradient is only used for pixels which are partially channel but have no upslope neighbours.

Method classify()

Split a catchment into a set hydrological response units (HRUs) according to any number of landscape layer cuts or burns

Usage

dynatopGIS$classify(cuts, burns = NULL)

Arguments

cuts

A named list of cuts of make to form the HRU. Names should correspond to raster layers in the project directory. Values should be numeric and define either the number of bands (single value) or breaks between band (multiple values)

burns

a vector of layer names which are to be used as burns

Details

This applies the given cuts to the supplied landscape layers to produce areal groupings of the catchment. Burns are added directly in the order they are given. Changing teh classification removes the current model ensuring the two are always syncronised.

Method get_class()

Get a classification layer of geographical information or a list of layer names

Usage

dynatopGIS$get_class(layer_name = character(0))

Arguments

layer_name

name of the layer give to the layer

Returns

a RasterLayer of the requested information if layer_name is given else a vector of layer names

Method plot_class()

Plot a classification layer

Usage

dynatopGIS$plot_class(layer_name = "final", add_channel = TRUE)

Arguments

layer_name

the name of layer to plot

add_channel

shouel the channel be added to the plot

Returns

a plot

Method create_model()

Compute a Dynamic TOPMODEL

Usage

dynatopGIS$create_model(band = 5, verbose = FALSE)

Arguments

band

How to cut the bands into the model, interpreted as for a cut in the classify method

verbose

print out additional diagnostic information

Method get_model()

Get the model is a form suitable for dynatop

Usage

dynatopGIS$get_model()

Returns

a Dynamic TOPMODE description suitable for the dyantop package

Method plot_model()

Plot the spatial properties of a model

Usage

dynatopGIS$plot_model(add_channel = TRUE)

Arguments

add_channel

should the channel be added to the plot

Method neighbour()

Return the index of neighbouring cells

Usage

dynatopGIS$neighbour(x)

Arguments

x

index of cells for which to find neighbours

Returns

a list of indexes of neighbours

Method get_class_method()

get the cuts and burns used to classify

Usage

dynatopGIS$get_class_method()

Returns

a list with two elements, cuts and burns

Method get_version()

get the version number

Usage

dynatopGIS$get_version()

Details

teh version number indicates the version of the algorithms within the object

Returns

a numeric version number

Method clone()

The objects of this class are cloneable with this method.

Usage

dynatopGIS$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.